pyproject.toml
[tool.poetry]
name = "screener"
version = "0.1.0"
description = "Modular quantitative research screener"
authors = ["James Liao"]
license = "Apache-2.0"
repository = "https://github.com/jamesjxliao/yass"
packages = [{include = "screener", from = "src"}]

[tool.poetry.scripts]
screener = "screener.cli:app"

[tool.poetry.dependencies]
python = "^3.12"
polars = "^1.0"
duckdb = "^1.0"
httpx = "^0.27"
pydantic = "^2.0"
pydantic-settings = "^2.0"
typer = "^0.12"
click = ">=8.1,<8.3"
pyyaml = "^6.0"
python-dateutil = "^2.9"
pyarrow = ">=15.0"
matplotlib = "^3.8"
streamlit = "^1.55.0"
alpaca-py = "^0.43.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ruff = "^0.5"

[tool.poetry.group.backtest]
optional = true

[tool.poetry.group.backtest.dependencies]
vectorbt = ">=0.26"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.ruff.lint.per-file-ignores]
"src/screener/evaluation/report.py" = ["N803", "N806"]
"tests/test_evaluation.py" = ["N806"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
